requirements.txt
beautifulsoup4==4.12.2
Flask==2.2.2
Flask_Cors==4.0.0
gpt4free==1.0.2
newspaper3k==0.2.8
Requests==2.31.0
gunicorn==20.1.0

